get_data: Added user headers for dta files
It is possible to add user header lines with the -H flag. The user header
is placed between the file header and the data, the offset is stored in
a new header field. Backwards compatibility since data offset is 0 for older
files, resulting in no user header available.

An assert has been added at the beginning of the program to ensure correct
binary header size (padding!), since I recorded quite some data with a
faulty padded header.