Modular data output for get_data program.
The main get_data routine now uses an abstract
data stream class to write the data to the hard drive.
The previously supported output modes are now implemented
as small, easy to replace and extend classes. So new
output formats do not require intervention
in the actual acquisition code.

In the mean time, the old binary format was
replaced by a new YAML-header based format, simplifying
the specification of arbitrary data in the header.