MultifileStream: Added error checking and sensible defaults
get_data crashed violently when no directory was specified,
because the default was empty, a slash was added and the
resulting directory was not writeable. As a consequence,
nullpointeres were followed...

Added some checks, and a check detecting an empty
directory name and use an autogenerated, datetime named
directory instead.